requirements.txt
numpy==1.16.4
pandas==1.0.3
matplotlib==3.1.1
tqdm==4.47.0
torchsummary==1.5.1
Pillow==6.1.0
scikit-learn==0.22.1
opencv-python==4.3.0.36
albumentations==0.4.5
torchviz==0.0.1
pyyaml==5.3.1
imutils==0.5.3
https://download.pytorch.org/whl/cu101/torch-1.4.0-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cu101/torchvision-0.5.0-cp36-cp36m-linux_x86_64.whl